@@ -1,86 +1,77 @@
 name: mi_utem
-description: Plataforma académica para estudiantes de la Universidad Tecnológica Metropolitana (UTEM)
+description: Plataforma académica para estudiantes de la Universidad Tecnológica
+  Metropolitana (UTEM)
+publish_to: none
 
 version: 2.4.0+38
 
 environment:
-  sdk: ">=2.10.4 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  flutter:
-    sdk: flutter
-  cupertino_icons: ^1.0.2
-  grpc: ^1.0.0
-  protobuf: ^0.13.12
-  flushbar: ^1.8.2
-  video_player: ^2.1.0
-  tinycolor: ^1.0.2
-  diagonal_scrollview: ^0.2.0
-  url_launcher: ^5.0.3
-  shared_preferences: ^0.4.3
-  font_awesome_flutter: ^8.5.0
-  flutter_masked_text: ^0.8.0
-  flutter_statusbar: ^0.0.1
-  dotted_border: ^1.0.2
-  barcode_flutter: ^1.0.2
-  flutter_keyboard_visibility: ^4.0.0
-  recase: ^2.0.1
-  flare_flutter: ^1.5.4
-  carousel_slider: ^3.0.0
-  dio: ^4.0.0
-  flutter_spinkit: ^4.1.2+1
-  dio_cache_interceptor: ^2.2.0
-  flutter_secure_storage: ^3.3.5
-  charts_flutter: ^0.10.0
-  mdi: ^3.0.0
-  sentry_flutter: ^4.1.0-nullsafety.0
-  firebase_core: ^1.0.1
-  firebase_analytics: ^7.1.1
-  firebase_remote_config: ^0.9.0-dev.1
-  firebase_in_app_messaging: ^0.4.0
-  firebase_messaging: ^9.1.2
-  cloud_firestore: ^1.0.5
-  html: ^0.14.0+4
-  gradient_widgets: ^0.5.1
-  timer_button: ^1.0.1
-  cached_network_image: ^2.5.0
-  circular_profile_avatar: ^1.0.6
-  photo_view: ^0.10.3
-  custom_refresh_indicator: ^0.9.0
-  in_app_review: ^1.0.4
-  simple_gesture_detector: ^0.1.6
-  extended_image: ^3.0.0
-  image_editor: ^0.7.3
-  image_editor_pro: ^1.0.9
-  image_picker: ^0.6.7+22
-  package_info: ^0.4.3+4
-  in_app_update: ^1.1.15
-  integration_test: ^1.0.2+2
-  badges: ^1.1.6
-  flutter_windowmanager: ^0.0.2
-  clipboard: ^0.1.2+8
-  flutter_markdown: ^0.6.1
-  flutter_local_notifications: ^5.0.0+1
-  awesome_notifications: 
+  awesome_notifications:
     git:
       url: git://github.com/aaqibism/awesome_notifications.git
+  badges: ^2.0.1
+  barcode_flutter: 
+    git:
+      url: git://github.com/venosyd/barcode.flutter.git
+  cached_network_image: ^3.1.0
+  carousel_slider: ^4.0.0
+  charts_flutter: ^0.11.0
+  circular_profile_avatar: ^2.0.5
+  clipboard: ^0.1.2+8
+  cloud_firestore: ^1.0.5
+  cupertino_icons: ^1.0.2
+  custom_refresh_indicator: ^1.0.0
+  dio: ^4.0.0
+  dio_cache_interceptor: ^3.0.1
+  dio_cache_interceptor_hive_store: ^3.0.1
+  dotted_border: ^2.0.0+1
+  extended_image: ^4.1.0
+  firebase_analytics: ^8.3.1
+  firebase_core: ^1.6.0
+  firebase_in_app_messaging: ^0.5.0+9
+  firebase_messaging: ^10.0.6
+  firebase_remote_config: ^0.10.0+5
+  flare_flutter: ^3.0.2
+  flutter:
+    sdk: flutter
+  flutter_keyboard_visibility: ^5.0.3
+  flutter_markdown: ^0.6.1
+  flutter_masked_text: 
+    git:
+      url: git://github.com/vitorhm/flutter-masked-text.git
+  flutter_secure_storage: ^4.2.1
+  flutter_spinkit: ^5.1.0
+  flutter_windowmanager: ^0.2.0
+  font_awesome_flutter: ^9.1.0
   get: ^4.1.4
+  gradient_widgets: ^0.6.0
+  html: ^0.15.0
+  image_editor: ^1.0.0
+  image_picker: ^0.8.4
+  in_app_review: ^2.0.3
+  in_app_update: ^2.0.0
+  mdi: ^5.0.0-nullsafety.0
+  package_info: ^2.0.2
+  permission_handler: ^8.1.4+2
+  photo_view: ^0.12.0
+  recase: ^4.0.0
   responsive_framework: ^0.1.0
-  permission_handler: ^7.1.0
+  sentry: ^6.0.0
+  shared_preferences: ^2.0.7
+  simple_gesture_detector: ^0.2.0
+  url_launcher: ^6.0.10
+  video_player: ^2.1.0
+  screenshot: ^1.2.3
+  image_gallery_saver: ^1.6.9
+  share_plus: ^2.1.4
+  path_provider: ^2.0.2
 
 dependency_overrides:
-  crypto: ^3.0.0
-  path_provider: ^2.0.1
-  http: ^0.13.0
-  package_info: ^2.0.0
-  plugin_platform_interface: ^2.0.0
-  http_parser: ^4.0.0
-  dio: ^4.0.0
-  integration_test: ^1.0.2+2
-  args: ^2.0.0
-  awesome_notifications: ^0.0.5+8
-  convert: ^3.0.0
-  permission_handler: ^7.1.0
+  image_picker: ^0.8.4
+  screenshot: ^1.2.3
 
 dev_dependencies:
   flutter_test:
