@@ -1,29 +1,31 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:mi_utem/models/evaluacion.dart';
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/themes/theme.dart';
 import 'package:recase/recase.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class Asignatura {
-   String nombre;
-   String codigo;
-   String tipoHora;
-   String estado;
-   String docente;
-   String seccion;
-   Color colorAsignatura;
-   List<Evaluacion> evaluaciones;
-   num notaExamen;
-   num notaPresentacion;
-   num notaFinal;
-   Asistencia asistencia;
-   List<Usuario> estudiantes;
-   String tipoAsignatura;
-   num intentos;
-   String horario;
-   String sala;
-   String tipoSala;
-
+  String? nombre;
+  String? codigo;
+  String? tipoHora;
+  String? estado;
+  String? docente;
+  String? seccion;
+  Color? colorAsignatura;
+  List<Evaluacion> evaluaciones;
+  num? notaExamen;
+  num? notaPresentacion;
+  num? notaFinal;
+  Asistencia? asistencia;
+  List<Usuario>? estudiantes;
+  String? tipoAsignatura;
+  num? intentos;
+  String? horario;
+  String? sala;
+  String? tipoSala;
 
   Asignatura({
     this.nombre,
@@ -43,19 +45,44 @@ class Asignatura {
     this.sala,
     this.horario,
     this.intentos,
-    this.tipoSala
+    this.tipoSala,
   });
 
-  factory Asignatura.fromJson(Map<String, dynamic> json) {
+  Future<void> addColor(List<Color> colores) async {
+    List<Color> coloresFiltrados = colores;
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    Map<String, dynamic> json = {};
+
+    if (prefs.containsKey('coloresAsignaturas') &&
+        prefs.getString('coloresAsignaturas') != null) {
+      json = jsonDecode(prefs.getString('coloresAsignaturas')!);
+    }
+
+    List<Color> coloresUsados = json.values.map((c) => Color(c)).toList();
+    coloresFiltrados.retainWhere((c) => !coloresUsados.contains(c));
+
+    if (!json.containsKey(this.codigo)) {
+      this.colorAsignatura = coloresFiltrados[0];
+      json[this.codigo!] = coloresFiltrados[0].value;
+      prefs.setString('coloresAsignaturas', jsonEncode(json));
+    } else {
+      this.colorAsignatura = Color(json[this.codigo!]!);
+    }
+  }
+
+  factory Asignatura.fromJson(Map<String, dynamic>? json) {
     if (json == null) {
       return Asignatura();
     }
-    return Asignatura(
+
+    var asignaturaParseada = Asignatura(
       codigo: json['codigo'],
       nombre: json['nombre'] != null ? ReCase(json['nombre']).titleCase : null,
-      tipoHora: json['tipoHora'] != null ? ReCase(json['tipoHora']).titleCase : null,
+      tipoHora:
+          json['tipoHora'] != null ? ReCase(json['tipoHora']).titleCase : null,
       estado: json['estado'] != null ? ReCase(json['estado']).titleCase : null,
-      docente: json['docente'] != null ? ReCase(json['docente']).titleCase : null,
+      docente:
+          json['docente'] != null ? ReCase(json['docente']).titleCase : null,
       seccion: json['seccion'],
       evaluaciones: Evaluacion.fromJsonList(json['evaluaciones']),
       notaFinal: json['notaFinal'],
@@ -63,12 +90,19 @@ class Asignatura {
       notaExamen: json['notaExamen'],
       estudiantes: Usuario.fromJsonList(json["estudiantes"]),
       asistencia: Asistencia.fromJson(json["asistencia"]),
-      tipoAsignatura: json['tipoAsignatura'] != null ? ReCase(json['tipoAsignatura']).titleCase : null,
+      tipoAsignatura: json['tipoAsignatura'] != null
+          ? ReCase(json['tipoAsignatura']).titleCase
+          : null,
       sala: json['sala'] != null ? ReCase(json['sala']).titleCase : null,
       horario: json['horario'],
       intentos: json['intentos'] ?? 0,
-      tipoSala: json['tipoSala'] != null ? ReCase(json['tipoSala']).titleCase : null,
+      tipoSala:
+          json['tipoSala'] != null ? ReCase(json['tipoSala']).titleCase : null,
     );
+
+    List<Color> colores = Colors.primaries.toList()..shuffle();
+    asignaturaParseada.addColor(colores);
+    return asignaturaParseada;
   }
 
   static List<Asignatura> fromJsonList(dynamic json) {
@@ -99,8 +133,8 @@ class Asignatura {
     double presentacion = 0;
     for (var evaluacion in evaluaciones) {
       if (evaluacion.nota != null && evaluacion.porcentaje != null) {
-        presentacion +=  evaluacion.nota * (evaluacion.porcentaje / 100);
-      }   
+        presentacion += evaluacion.nota! * (evaluacion.porcentaje! / 100);
+      }
     }
 
     return presentacion;
@@ -117,7 +151,8 @@ class Asignatura {
 
   bool get puedeDarExamen {
     if (estanTodasLasNotas) {
-      return 2.95 <= this.notaPresentacionCalculada && this.notaPresentacionCalculada < 3.95;
+      return 2.95 <= this.notaPresentacionCalculada &&
+          this.notaPresentacionCalculada < 3.95;
     }
     return false;
   }
@@ -158,8 +193,7 @@ class Asignatura {
 
   num get notaFinalCalculada {
     if (puedeDarExamen && notaExamen != null) {
-      return notaPresentacionCalculada * 0.6 + notaExamen * 0.4;
-     
+      return notaPresentacionCalculada * 0.6 + notaExamen! * 0.4;
     } else {
       return notaPresentacionCalculada;
     }
@@ -179,7 +213,7 @@ class Asistencia {
     this.sinRegistro = 0,
   });
 
-  factory Asistencia.fromJson(Map<String, dynamic> json) {
+  factory Asistencia.fromJson(Map<String, dynamic>? json) {
     if (json == null) {
       return Asistencia();
     }
@@ -201,6 +235,4 @@ class Asistencia {
     }
     return list;
   }
-
-
 }
