@@ -2,12 +2,10 @@ import 'dart:core';
 
 import 'package:clipboard/clipboard.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mi_utem/models/asignatura.dart';
 import 'package:mi_utem/models/usuario.dart';
-import 'package:mi_utem/services/autenticacion_service.dart';
 import 'package:mi_utem/services/docentes_service.dart';
 import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/services/review_service.dart';
@@ -17,22 +15,22 @@ import 'package:mi_utem/widgets/image_view_screen.dart';
 import 'package:mi_utem/widgets/loading_dialog.dart';
 import 'package:mi_utem/widgets/loading_indicator.dart';
 import 'package:mi_utem/widgets/profile_photo.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class UsuarioScreen extends StatefulWidget {
   final int tipo;
-  final Map<String, dynamic> query;
-  final Asignatura asignatura;
-  UsuarioScreen({Key key, this.tipo = 0, this.query, this.asignatura}) : super(key: key);
+  final Map<String, dynamic>? query;
+  final Asignatura? asignatura;
+  UsuarioScreen({Key? key, this.tipo = 0, this.query, this.asignatura})
+      : super(key: key);
 
   @override
   _UsuarioScreenState createState() => _UsuarioScreenState();
 }
 
 class _UsuarioScreenState extends State<UsuarioScreen> {
-  Future<Usuario> _usuarioFuture;
-  Usuario _usuario;
+  Future<Usuario>? _usuarioFuture;
+  Usuario? _usuario;
 
   @override
   void initState() {
@@ -48,11 +46,15 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
       if (widget.tipo == 2) {
         print(widget.query);
         if (widget.asignatura == null) {
-          usuario = await DocentesService.traerUnDocente(widget.query["nombre"]);
+          usuario =
+              await DocentesService.traerUnDocente(widget.query!["nombre"]);
         } else {
-          usuario = await DocentesService.asignarUnDocente(widget.query["nombre"], widget.asignatura.codigo, widget.asignatura.nombre);
+          usuario = await DocentesService.asignarUnDocente(
+              widget.query!["nombre"],
+              widget.asignatura!.codigo,
+              widget.asignatura!.nombre);
         }
-        
+
         setState(() {
           _usuario = usuario;
         });
@@ -83,21 +85,25 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
         _usuario = usuario;
       });
 
-      return usuario;
+      return;
     } catch (e) {
       Get.back();
       print("Error cambiando la imagen ${e.toString()}");
-      Flushbar(
-        message: "No se pudo cambiar la foto",
-        duration: Duration(seconds: 3),
-      )..show(context);
+      Get.snackbar(
+        "Error",
+        "No se pudo cambiar la foto",
+        colorText: Colors.white,
+        backgroundColor: Get.theme.primaryColor,
+        snackPosition: SnackPosition.BOTTOM,
+        margin: EdgeInsets.all(20),
+      );
     }
   }
 
   List<Widget> get _datosPersonales {
     List<Widget> lista = [];
     if (_usuario != null) {
-      if (_usuario.nombre != null && _usuario.nombre.isNotEmpty) {
+      if (_usuario!.nombre != null && _usuario!.nombre!.isNotEmpty) {
         lista.add(ListTile(
           title: Text(
             "Nombre",
@@ -106,7 +112,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
             ),
           ),
           subtitle: Text(
-            _usuario.nombreCompleto,
+            _usuario!.nombreCompleto!,
             style: TextStyle(
               color: Colors.grey[900],
               fontSize: 18,
@@ -114,7 +120,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
           ),
         ));
       } else {
-        if (_usuario.nombres != null && _usuario.nombres.isNotEmpty) {
+        if (_usuario!.nombres != null && _usuario!.nombres!.isNotEmpty) {
           lista.add(
             ListTile(
               title: Text(
@@ -124,7 +130,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
                 ),
               ),
               subtitle: Text(
-                _usuario.nombres,
+                _usuario!.nombres!,
                 style: TextStyle(
                   color: Colors.grey[900],
                   fontSize: 18,
@@ -134,7 +140,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
           );
         }
 
-        if (_usuario.apellidos != null && _usuario.apellidos.isNotEmpty) {
+        if (_usuario!.apellidos != null && _usuario!.apellidos!.isNotEmpty) {
           lista.add(Divider(height: 1));
           lista.add(
             ListTile(
@@ -145,7 +151,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
                 ),
               ),
               subtitle: Text(
-                _usuario.apellidos,
+                _usuario!.apellidos!,
                 style: TextStyle(
                   color: Colors.grey[900],
                   fontSize: 18,
@@ -156,7 +162,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
         }
       }
 
-      if (_usuario.correo != null && _usuario.correo.isNotEmpty) {
+      if (_usuario!.correo != null && _usuario!.correo!.isNotEmpty) {
         lista.add(Divider(height: 1));
         lista.add(ListTile(
           title: Text(
@@ -165,18 +171,26 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
               color: Colors.grey,
             ),
           ),
-          onLongPress: widget.tipo != 0 ? () async {
-            await FlutterClipboard.copy(_usuario.correo);
-            Flushbar(
-              message: "Correo copiado al portapapeles",
-              duration: Duration(seconds: 3),
-            )..show(context);
-          } : null,
-          onTap: widget.tipo != 0 ? () async {
-            await launch("mailto:${_usuario.correo}");
-          } : null,
+          onLongPress: widget.tipo != 0
+              ? () async {
+                  await FlutterClipboard.copy(_usuario!.correo!);
+                  Get.snackbar(
+                    "¡Copiado!",
+                    "Correo copiado al portapapeles",
+                    colorText: Colors.white,
+                    backgroundColor: Get.theme.primaryColor,
+                    snackPosition: SnackPosition.BOTTOM,
+                    margin: EdgeInsets.all(20),
+                  );
+                }
+              : null,
+          onTap: widget.tipo != 0
+              ? () async {
+                  await launch("mailto:${_usuario!.correo}");
+                }
+              : null,
           subtitle: Text(
-            _usuario.correo,
+            _usuario!.correo!,
             style: TextStyle(
               color: Colors.grey[900],
               fontSize: 18,
@@ -185,7 +199,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
         ));
       }
 
-      if (widget.tipo == 0 && _usuario.rut != null) {
+      if (widget.tipo == 0 && _usuario!.rut != null) {
         lista.add(Divider(height: 1));
         lista.add(ListTile(
           title: Text(
@@ -195,7 +209,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
             ),
           ),
           subtitle: Text(
-            _usuario.rut.formateado(true),
+            _usuario!.rut!.formateado(true),
             style: TextStyle(
               color: Colors.grey[900],
               fontSize: 18,
@@ -212,7 +226,7 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: CustomAppBar(
-        title: Text(widget.tipo == 0 ? "Perfil" : widget.query["nombre"]),
+        title: Text(widget.tipo == 0 ? "Perfil" : widget.query!["nombre"]),
       ),
       body: FutureBuilder(
         future: _usuarioFuture,
@@ -242,20 +256,19 @@ class _UsuarioScreenState extends State<UsuarioScreen> {
                     ),
                     Center(
                       child: ProfilePhoto(
-                        usuario: _usuario,
-                        radius: 60,
-                        editable: widget.tipo == 0,
-                        onImage: widget.tipo == 0
-                            ? (image) {
-                                _changeFoto(image);
-                              }
-                            : null,
-                        onImageTap: (context, imageProvider) {
-                          Get.to(
-                            ImageViewScreen(imagen: imageProvider),
-                          );
-                        }
-                      ),
+                          usuario: _usuario,
+                          radius: 60,
+                          editable: widget.tipo == 0,
+                          onImage: widget.tipo == 0
+                              ? (image) {
+                                  _changeFoto(image);
+                                }
+                              : null,
+                          onImageTap: (context, imageProvider) {
+                            Get.to(
+                              ImageViewScreen(imagen: imageProvider),
+                            );
+                          }),
                     ),
                   ],
                 ),
