@@ -9,7 +9,9 @@ import 'package:firebase_analytics_web/firebase_analytics_web.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
 import 'package:firebase_messaging_web/firebase_messaging_web.dart';
 import 'package:flutter_keyboard_visibility_web/flutter_keyboard_visibility_web.dart';
-import 'package:sentry_flutter/sentry_flutter_web.dart';
+import 'package:image_picker_for_web/image_picker_for_web.dart';
+import 'package:share_plus_web/share_plus_web.dart';
+import 'package:shared_preferences_web/shared_preferences_web.dart';
 import 'package:url_launcher_web/url_launcher_web.dart';
 import 'package:video_player_web/video_player_web.dart';
 
@@ -22,7 +24,9 @@ void registerPlugins(Registrar registrar) {
   FirebaseCoreWeb.registerWith(registrar);
   FirebaseMessagingWeb.registerWith(registrar);
   FlutterKeyboardVisibilityPlugin.registerWith(registrar);
-  SentryFlutterWeb.registerWith(registrar);
+  ImagePickerPlugin.registerWith(registrar);
+  SharePlusPlugin.registerWith(registrar);
+  SharedPreferencesPlugin.registerWith(registrar);
   UrlLauncherPlugin.registerWith(registrar);
   VideoPlayerPlugin.registerWith(registrar);
   registrar.registerMessageHandler();
