@@ -1,11 +1,7 @@
-import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
-import 'package:mi_utem/models/asignatura.dart';
 import 'package:mi_utem/models/horario.dart';
-import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/utils/dio_miutem_client.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class HorarioService {
   static final Dio _dio = DioMiUtemClient.authDio;
@@ -18,16 +14,16 @@ class HorarioService {
         uri,
         options: DioMiUtemClient.cacheOptions
             .copyWith(
-                maxStale: Duration(hours: 3),
+                maxStale: Duration(days: 30),
                 policy: refresh ? CachePolicy.refresh : CachePolicy.forceCache)
             .toOptions(),
       );
 
       Horario horario = Horario.fromJson(response.data);
 
-      for (var dia in horario.horario) {
+      for (var dia in horario.horario!) {
         for (var bloque in dia) {
-          print("HorarioService bloque ${bloque?.codigo}");
+          print("HorarioService bloque ${bloque.codigo}");
         }
       }
 
