@@ -1,18 +1,11 @@
-import 'dart:convert';
+import 'dart:developer';
+
 import 'package:awesome_notifications/awesome_notifications.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:dio/dio.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:in_app_review/in_app_review.dart';
-import 'package:mi_utem/models/rut.dart';
-import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/themes/theme.dart';
-import 'package:mi_utem/utils/dio_miutem_client.dart';
 import 'package:mi_utem/widgets/custom_alert_dialog.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class NotificationsService {
   static AwesomeNotifications get notification => AwesomeNotifications();
@@ -20,14 +13,27 @@ class NotificationsService {
   static Future initialize() async {
     try {
       FirebaseMessaging.onBackgroundMessage(_onFCM);
-      await notification.initialize(null, [
-        NotificationChannel(
-            channelKey: 'basic_channel',
-            channelName: 'Basic notifications',
-            channelDescription: 'Notification channel for basic tests',
-            defaultColor: MainTheme.primaryColor,
-            ledColor: MainTheme.primaryColor)
-      ]);
+      FirebaseMessaging.onMessage.listen((RemoteMessage message) async {
+        await notification.createNotification(
+          content: NotificationContent(
+            id: 0,
+            channelKey: "basic_channel",
+            title: message.notification?.title ?? "",
+            body: message.notification?.body ?? "",
+          ),
+        );
+      });
+      await notification.initialize(
+        null,
+        [
+          NotificationChannel(
+              channelKey: 'basic_channel',
+              channelName: 'Basic notifications',
+              channelDescription: 'Notification channel for basic tests',
+              defaultColor: MainTheme.primaryColor,
+              ledColor: MainTheme.primaryColor)
+        ],
+      );
       NotificationsService.notification.actionStream
           .listen((receivedNotification) {});
 
@@ -54,7 +60,7 @@ class NotificationsService {
                 dismissedAction.body ??
                 dismissedAction.id.toString()));
       });
-      String fcmToken = await FirebaseMessaging.instance.getToken();
+      String? fcmToken = await FirebaseMessaging.instance.getToken();
       print("Firebase token: $fcmToken");
     } catch (e) {
       print('Error ${e.toString()}');
@@ -63,21 +69,15 @@ class NotificationsService {
 
   static Future<dynamic> _onFCM(RemoteMessage message) async {
     WidgetsFlutterBinding.ensureInitialized();
-    notification.createNotification(
+    log("Notification received");
+    /* notification.createNotification(
       content: NotificationContent(
         channelKey: "basic_channel",
         id: 1,
-        title: message.notification?.title ?? "New Notification",
+        title: message.notification?.title ?? "",
         body: message.notification?.body ?? "",
       ),
-      actionButtons: [
-        NotificationActionButton(
-          key: 'Reply',
-          label: 'Reply',
-          buttonType: ActionButtonType.InputField,
-        ),
-      ],
-    );
+    ); */
   }
 
   static Future<bool> requestUserPermissionIfNecessary() async {
@@ -87,20 +87,22 @@ class NotificationsService {
         CustomAlertDialog(
           titulo: "Activa las notificaciones",
           emoji: "🔔",
-          descripcion: "Necesitamos tu permiso para poder enviarte notificaciones. Nada de spam, lo prometemos.",
-            onCancelar: () async {
-              bool isAllowed = await notification.isNotificationAllowed();
-              Get.back(result: isAllowed);
-              Get.back(result: isAllowed);
-            },
-            onConfirmar: () async {
-              await notification.requestPermissionToSendNotifications();
-              bool isAllowed = await notification.isNotificationAllowed();
-              Get.back(result: isAllowed);
-              Get.back(result: isAllowed);
-            },
-            cancelarTextoBoton: "No permitir",
-            confirmarTextoBoton: "Permitir",),
+          descripcion:
+              "Necesitamos tu permiso para poder enviarte notificaciones. Nada de spam, lo prometemos.",
+          onCancelar: () async {
+            bool isAllowed = await notification.isNotificationAllowed();
+            Get.back(result: isAllowed);
+            Get.back(result: isAllowed);
+          },
+          onConfirmar: () async {
+            await notification.requestPermissionToSendNotifications();
+            bool isAllowed = await notification.isNotificationAllowed();
+            Get.back(result: isAllowed);
+            Get.back(result: isAllowed);
+          },
+          cancelarTextoBoton: "No permitir",
+          confirmarTextoBoton: "Permitir",
+        ),
       );
     }
     return isAllowed;
